services/demo-web/requirements.txt
flask
requests
google-cloud-pubsub
pycrypto
redis
geoip2
